+++ src/main/java/broker/ForexBroker.java
@@ -26,6 +26,6 @@ public interface ForexBroker {
 
     void openPosition(ForexTrader trader, OpenPositionRequest request) throws Exception;
 
-    void closePosition(ForexTrader trader, ForexPosition position, @Nullable Double limit) throws Exception;
+    void closePosition(ForexTrader trader, ForexPosition position, @Nullable Long limit) throws Exception;
 
 }
